feat: add link to heading h1, h2
